backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.0
gunicorn>=21.0
dj-database-url>=2.2
psycopg2-binary>=2.9
mysqlclient>=2.2
